+++ CHANGELOG.md
@@ -23,6 +23,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [1.1.1] - 2025-09-22
+
+### Fixed
+
+- apply odk variable filter rule
+
 ## [1.1.0] - 2025-09-11
 
 ### Added